Extracted DeploymentManagerDispatcher, ProcessStateService and CustomActionInvokerService
