+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/process/deployment/ManagementActor.scala
@@ -2,28 +2,23 @@ package pl.touk.nussknacker.ui.process.deployment
 
 import akka.actor.{ActorRefFactory, Props, Status}
 import com.typesafe.scalalogging.LazyLogging
-import pl.touk.nussknacker.engine
 import pl.touk.nussknacker.engine.api.deployment.ProcessActionType.ProcessActionType
 import pl.touk.nussknacker.engine.api.deployment._
-import pl.touk.nussknacker.engine.api.deployment.simple.{SimpleProcessStateDefinitionManager, SimpleStateStatus}
-import pl.touk.nussknacker.engine.api.process.{ProcessId, ProcessName, VersionId}
+import pl.touk.nussknacker.engine.api.deployment.simple.SimpleStateStatus
+import pl.touk.nussknacker.engine.api.process.{ProcessName, VersionId}
 import pl.touk.nussknacker.engine.api.test.ScenarioTestData
 import pl.touk.nussknacker.engine.canonicalgraph.CanonicalProcess
 import pl.touk.nussknacker.engine.deployment.{User => ManagerUser}
 import pl.touk.nussknacker.restmodel.process.ProcessIdWithName
 import pl.touk.nussknacker.ui.EspError
-import pl.touk.nussknacker.ui.api.ListenerApiUser
 import pl.touk.nussknacker.ui.db.entity.ProcessActionEntityData
-import pl.touk.nussknacker.ui.listener.{User => ListenerUser}
 import pl.touk.nussknacker.ui.process.processingtypedata.ProcessingTypeDataProvider
-import pl.touk.nussknacker.ui.process.repository.ProcessDBQueryRepository.ProcessNotFoundError
 import pl.touk.nussknacker.ui.process.repository.{DeploymentComment, FetchingProcessRepository}
 import pl.touk.nussknacker.ui.security.api.LoggedUser
 import pl.touk.nussknacker.ui.util.FailurePropagatingActor
 
 import java.time.Instant
 import scala.concurrent.{ExecutionContext, Future}
-import scala.util.control.NonFatal
 import scala.util.{Failure, Success}
 
 object ManagementActor {
@@ -32,12 +27,15 @@ object ManagementActor {
             scenarioResolver: ScenarioResolver,
             deploymentService: DeploymentService)
            (implicit context: ActorRefFactory): Props = {
+    val dispatcher = new DeploymentManagerDispatcher(managers, processRepository)
+    val processStateService = new ProcessStateService(processRepository, dispatcher, deploymentService)
     Props(
       classOf[ManagementActor],
-      processRepository,
-      new DeploymentManagerDispatcher(managers, processRepository),
+      dispatcher,
       scenarioResolver,
-      deploymentService)
+      deploymentService,
+      new CustomActionInvokerService(processRepository, dispatcher, processStateService),
+      processStateService)
   }
 }
 
@@ -46,13 +44,13 @@ object ManagementActor {
 // - protecting that there won't be more than one scenario being deployed simultaneously
 // - being able to check status of asynchronous perform deploy operation
 // Already extracted is only DeploymentService - see docs there, but should be extracted more classes e.g.:
-// - translating (ProcessState from DeploymentManager and historical context of deployment/cancel actions) to user-friendly ProcessStatus
 // - subprocess resolution should be a part of kind of ResolvedProcessRepository
 // - (maybe) some kind of facade spinning all this things together and not being an actor e.g. ScenarioManagementFacade
-class ManagementActor(processRepository: FetchingProcessRepository[Future],
-                      dispatcher: DeploymentManagerDispatcher,
+class ManagementActor(dispatcher: DeploymentManagerDispatcher,
                       scenarioResolver: ScenarioResolver,
-                      deploymentService: DeploymentService) extends FailurePropagatingActor with LazyLogging {
+                      deploymentService: DeploymentService,
+                      customActionInvokerService: CustomActionInvokerService,
+                      processStateService: ProcessStateService) extends FailurePropagatingActor with LazyLogging {
 
   private var beingDeployed = Map[ProcessName, DeployInfo]()
 
@@ -87,9 +85,9 @@ class ManagementActor(processRepository: FetchingProcessRepository[Future],
       } yield manager.processStateDefinitionManager.processState(SimpleStateStatus.DuringDeploy)
       reply(processStatus)
     case CheckStatus(id, user) =>
-      reply(getProcessStatus(id)(user))
+      implicit val loggedUser: LoggedUser = user
+      reply(processStateService.getProcessState(id))
     case DeploymentActionFinished(process, user, _) =>
-      implicit val listenerUser: ListenerUser = ListenerApiUser(user)
       beingDeployed -= process.name
     case Test(id, canonicalProcess, category, scenarioTestData, user, encoder) =>
       ensureNoDeploymentRunning {
@@ -105,54 +103,16 @@ class ManagementActor(processRepository: FetchingProcessRepository[Future],
       reply(Future.successful(DeploymentStatusResponse(beingDeployed)))
 
     case CustomAction(actionName, id, user, params) =>
-      implicit val loggedUser: LoggedUser = user
       // TODO: Currently we're treating all custom actions as deployment actions; i.e. they can't be invoked if there is some deployment in progress
       ensureNoDeploymentRunning {
-        val maybeProcess = processRepository.fetchLatestProcessDetailsForProcessId[CanonicalProcess](id.id)
-        val res: Future[Either[CustomActionError, CustomActionResult]] = maybeProcess.flatMap {
-          case Some(process) =>
-            val actionReq = engine.api.deployment.CustomActionRequest(
-              name = actionName,
-              processVersion = process.toEngineProcessVersion,
-              user = toManagerUser(user),
-              params = params)
-            dispatcher.deploymentManager(id.id).flatMap { manager =>
-              manager.customActions.find(_.name == actionName) match {
-                case Some(customAction) =>
-                  getProcessStatus(id).flatMap(status => {
-                    if (customAction.allowedStateStatusNames.contains(status.status.name)) {
-                      manager.invokeCustomAction(actionReq, process.json)
-                    } else
-                      Future(Left(CustomActionInvalidStatus(actionReq, status.status.name)))
-                  })
-                case None =>
-                  Future(Left(CustomActionNonExisting(actionReq)))
-              }
-            }
-          case None =>
-            Future.failed(ProcessNotFoundError(id.id.value.toString))
-        }
+        implicit val loggedUser: LoggedUser = user
+        val res = customActionInvokerService.invokeCustomAction(actionName, id, params)
         reply(res)
       }
   }
 
-  private def getProcessStatus(processIdWithName: ProcessIdWithName)(implicit user: LoggedUser): Future[ProcessState] =
-    for {
-      actions <- processRepository.fetchProcessActions(processIdWithName.id)
-      manager <- dispatcher.deploymentManager(processIdWithName.id)
-      state <- findJobState(manager, processIdWithName)
-      _ <- deploymentService.handleFinishedProcess(processIdWithName, state)
-    } yield ObsoleteStateDetector.handleObsoleteStatus(state, actions.headOption)
-
-  private def findJobState(deploymentManager: DeploymentManager, processIdWithName: ProcessIdWithName)(implicit user: LoggedUser): Future[Option[ProcessState]] =
-    deploymentManager.findJobStatus(processIdWithName.name).recover {
-      case NonFatal(e) =>
-        logger.warn(s"Failed to get status of ${processIdWithName}: ${e.getMessage}", e)
-        Some(SimpleProcessStateDefinitionManager.processState(SimpleStateStatus.FailedToGet))
-    }
-
   private def handleDeploymentAction(id: ProcessIdWithName, user: LoggedUser, action: DeploymentActionType, deploymentComment: Option[DeploymentComment],
-                                 actionFuture: Future[ProcessActionEntityData]): Unit = {
+                                     actionFuture: Future[ProcessActionEntityData]): Unit = {
     beingDeployed += id.name -> DeployInfo(user.username, System.currentTimeMillis(), action)
     actionFuture.onComplete {
       case Success(details) => self ! DeploymentActionFinished(id, user, Right(DeploymentDetails(details.processVersionId, deploymentComment,details.performedAtTime, details.action)))
